Complete primes

+++ problems/convoy/main.cpp
@@ -2,12 +2,12 @@
 #include <queue>
 using namespace std;
 
-long long n, k, *t, *d, nd;
+int n, k, *t, *d, nd;
 
-bool verify(long long tlimit) {
-    long long num = 0;
+bool verify(int tlimit) {
+    int num = 0;
     for (int i = 0; i < nd; ++i) {
-        long long r = tlimit / d[i];
+        int r = tlimit / d[i];
         if (r > 0) {
             num += 1 + (r + 1) / 2 * 4;
         }
@@ -22,8 +22,8 @@ int main(int argc, char const *argv[]) {
 
     cin >> n >> k;
     nd = min(n, k);
-    t = new long long[n];
-    d = new long long[nd];
+    t = new int[n];
+    d = new int[nd];
 
     priority_queue<int, vector<int>, greater<int>> q;
 
@@ -37,11 +37,11 @@ int main(int argc, char const *argv[]) {
         q.pop();
     }
 
-    long long r = (2 * ((n - 1) / 4 + 1) - 1) * d[0], l = 0;
+    int r = (2 * ((n - 1) / 4 + 1) - 1) * d[0], l = 0;
 
     // at any time: v(r) = true but v(l) = false
     while (r - l > 1) {
-        long long mid = l + (r - l) / 2;
+        int mid = l + (r - l) / 2;
         if (verify(mid)) {
             r = mid;
         } else {
